[ALPS04016653] DISP: fix display build error in 32bit
[Detail]
fix display linker error in kernel 32bit project

MTK-Commit-Id: 9894dcebc164dd658657764810411575531ad7ab

CR-Id: ALPS04016653
Feature: Display Driver (DSI Interface)
(cherry-pick from 3e0aaf43613d46eb11bafd19090dd80b1e0d1d41)

+++ drivers/misc/mediatek/video/mt6771/videox/debug.c
@@ -83,7 +83,7 @@ int hrt_show_flag;
 int layer_en_num;
 int layer_en_num_flag;
 uint64_t layer_size_high;
-int layer_size_low;
+uint64_t layer_size_low;
 int layer_size_flag;
 
 /* path mode */
